Split internal Oxlint rules by module
The plugin is loaded directly from TypeScript source, so Node requires explicit `.ts` extensions for relative modules. The root TypeScript project rejected those specifiers, forcing every rule to remain coupled to the plugin entry point.

Give the tool its own TypeScript project so that exception stays local while the workspace still typechecks it explicitly. Individual rules can now live in focused modules as the internal rule set grows.
