+++ tools/oxlint-rules/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "allowImportingTsExtensions": true,
+    "tsBuildInfoFile": "../../node_modules/.cache/tsc/oxlint-rules.tsbuildinfo"
+  },
+  "include": ["*.ts", "src/**/*", "test/**/*"],
+  "exclude": ["coverage", "node_modules"]
+}
